Task 2/requirements.txt
streamlit
requests
python-dotenv
pandas
plotly